feat: Add native Moza KS wheel support (64-bit button state and new Extra bindings)
- Extended direct_wheel input pipeline from 32-bit to 64-bit buttons to support Moza bases reporting high button indices (e.g. KS wheel up to 38).
- Registered 'MOZA' base in device registry with proper mappings.
- Renamed Mod Settings UI category from 'G-series Wheel' to 'Wheel Mod' to reflect vendor-agnostic status.
- Added 10 Extra Button slots to Mod Settings for mapping advanced wheel inputs (N, WIP, FL, CAM, S1, S2, etc.).
- Removed legacy patched mod_settings.dll bundling to fix Vortex installation strictness.

+++ direct_wheel/src/sources.h
@@ -23,7 +23,7 @@ namespace direct_wheel::sources
 
     struct DigitalFrame
     {
-        uint32_t buttons = 0;       // bit per DInput button, low 32
+        uint64_t buttons = 0;       // bit per DInput button, low 64
         uint16_t pov     = 0xFFFF;  // POV[0] raw value, 0xFFFF = center
     };
